Replace localstack with moto-backed awsmock component
localstack/localstack:latest broke CI — all PRs fail with exit code 137
(OOM kill) during test:coverAll because the container fails to start.

Replace with motoserver/moto via a new generic awsmock package:
- New docker/component/awsmock package backed by motoserver/moto
- Deprecate docker/component/localstack (delegates to awsmock)
- WithServices becomes a no-op (moto exposes all services by default)
- Health check: /moto-api/ on port 5000

Tested locally: test:coverAll passes in ~52s.
